+++ src/lightly_train/_train_task_state.py
@@ -35,16 +35,3 @@ class CheckpointDict(TypedDict):
     # Used to reconstruct the model after training.
     model_class_path: str
     model_init_args: dict[str, Any]
-
-
-# Checkpoint as saved in out/checkpoints/last.ckpt
-class TrainCheckpoint(CheckpointDict):
-    optimizer_state_dict: dict[str, Any]
-    scheduler_state_dict: dict[str, Any]
-    train_dataloader: DataLoader[TaskBatch]
-    step: int
-
-
-# Checkpoint as saved in out/exported/last.pt
-class ExportedCheckpoint(CheckpointDict):
-    pass